+++ circuitx/gesture-navigation/src/iosMain/kotlin/com/slack/circuitx/gesturenavigation/iOSPredictiveBackNavigationDecoration.kt
@@ -44,14 +44,12 @@ private val End: (Int) -> Int = { it }
  * A factory that creates an [IOSPredictiveBackNavDecorator] for iOS predictive back navigation.
  *
  * @param fallback The [AnimatedNavDecorator.Factory] to use when predictive back is not supported.
- * @param onBackInvoked A callback to be invoked when a back gesture is performed.
  * @return An [AnimatedNavDecorator.Factory] that provides iOS predictive back navigation.
  */
 public actual fun GestureNavigationDecorationFactory(
-  fallback: AnimatedNavDecorator.Factory,
-  onBackInvoked: () -> Unit,
+  fallback: AnimatedNavDecorator.Factory
 ): AnimatedNavDecorator.Factory {
-  return IOSPredictiveBackNavDecorator.Factory(onBackInvoked = onBackInvoked)
+  return IOSPredictiveBackNavDecorator.Factory()
 }
 
 /**
@@ -61,12 +59,10 @@ public actual fun GestureNavigationDecorationFactory(
  *
  * @property enterOffsetFraction The fraction (from 0f to 1f) of the entering content's width which
  *   the content starts from. Defaults to 0.25f (25%).
- * @param onBackInvoked A callback to be invoked when a back gesture is performed.
  */
 internal class IOSPredictiveBackNavDecorator<T : NavArgument>(
-  private val enterOffsetFraction: Float = 0.25f,
-  onBackInvoked: () -> Unit,
-) : PredictiveBackNavigationDecorator<T>(onBackInvoked) {
+  private val enterOffsetFraction: Float = 0.25f
+) : PredictiveBackNavigationDecorator<T>() {
 
   // Track popped zIndex so screens are layered correctly
   private var zIndexDepth = 0f
@@ -117,15 +113,10 @@ internal class IOSPredictiveBackNavDecorator<T : NavArgument>(
     }
   }
 
-  internal class Factory(
-    private val enterOffsetFraction: Float = 0.25f,
-    private val onBackInvoked: () -> Unit,
-  ) : AnimatedNavDecorator.Factory {
+  internal class Factory(private val enterOffsetFraction: Float = 0.25f) :
+    AnimatedNavDecorator.Factory {
     override fun <T : NavArgument> create(): AnimatedNavDecorator<T, *> {
-      return IOSPredictiveBackNavDecorator(
-        enterOffsetFraction = enterOffsetFraction,
-        onBackInvoked = onBackInvoked,
-      )
+      return IOSPredictiveBackNavDecorator(enterOffsetFraction = enterOffsetFraction)
     }
   }
 }