+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Unreleased
 
 ### Changes
 
+- **Breaking**: Moved `NavDecoration` from `backstack` to `circuit-foundation` and added a `Navigator` parameter to `DecoratedContent`. `AnimatedNavDecorator` now has an `updateNavigator` method to receive the new `Navigator` parameter. This gives decorations direct access to the correct Navigator for handling back gestures.
 - Update to Kotlin `2.3.20`.
 - Remove deprecated X64 Apple targets.
 - `SaveableNavStack` & `SaveableBackStack` - Made the `Saver`s public apis